refactor(frontend): decompose Message.jsx into focused modules
Message.jsx had grown to 1,396 lines and mixed rendering, RAG citation
post-processing, markdown/highlight.js setup, clipboard handling, tool
argument/result shaping, and two sibling components into one file. This
change extracts cohesive slices into siblings without touching behavior:

- utils/markdownRenderer.js  - marked renderer, hljs language registration,
                               DOMPurify config
- utils/ragCitations.js      - extractSourceLabels, processCitationBadges,
                               processReferencesSection
- utils/messageContent.js    - processMessageContent, file refs, bullet norm
- utils/clipboard.js         - copyCodeBlock, copyMessageContent
- utils/toolResultUtils.js   - filterArgumentsForDisplay, processToolResult,
                               downloadReturnedFile
- components/ToolApprovalMessage.jsx, components/ToolElapsedTime.jsx

Message.jsx is now 524 lines, under the AGENTS.md 400-line target for most
of the file's body (the bulk that remains is the tool_call result renderer,
which is cohesive and harder to split safely).

rag-citation-rendering.test.js previously duplicated the helpers it was
testing; it now imports from utils/ragCitations.js so future changes cannot
drift from the tested contract.

All 372 frontend tests pass. Lint is clean. Build succeeds.
